Added new Functionality like Slideshow.Make changes in folder structure.Css changes .Added images in public folder.
