+++ ChildHome/package.json
@@ -7,9 +7,12 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
     "config": "^3.3.12",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.10.9",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.4.0",
     "react-router-dom": "^7.1.3",
     "react-scripts": "5.0.1",
     "react-toastify": "^10.0.6",
