+++ ChildHome/src/Components/Admin/admin_Sidebar.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const admin_Sidebar = () => {
+  return <div>admin_Sidebar</div>;
+};
+
+export default admin_Sidebar;
