+++ ChildHome/src/Components/Donation/RazorpayPayment.css
@@ -1,0 +1,33 @@
+.payment-card {
+    border-radius: 15px;
+    background: linear-gradient(135deg, #f6d365 0%, #fda085 100%);
+    color: #fff;
+  }
+  
+  .payment-card .card-title {
+    font-size: 2.5rem;
+    font-weight: bold;
+  }
+  
+  .payment-card .form-control {
+    border-radius: 10px;
+    border: none;
+    padding: 10px;
+  }
+  
+  .payment-card .pay-button {
+    background-color: #ff6f61;
+    border: none;
+    border-radius: 10px;
+    padding: 10px;
+    font-size: 1.2rem;
+    transition: background-color 0.3s ease;
+  }
+  
+  .payment-card .pay-button:hover {
+    background-color: #ff3b2f;
+  }
+  
+  .payment-card .text-muted {
+    color: #e0e0e0 !important;
+  }
