--- ChildHome/src/Components/Login/ContactUs.jsx
@@ -1,65 +1,0 @@
-import React, { useState } from 'react';
-
-const ContactUs = () => {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [message, setMessage] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    // Handle form submission here (like sending the data to an API)
-    console.log({ name, email, message });
-  };
-
-  return (
-    <div className="container my-5">
-      <div className="text-center mb-4">
-        <h1 className="text-primary">Contact Us</h1>
-        <p className="lead text-muted">We'd love to hear from you! Please fill out the form below.</p>
-      </div>
-      <form onSubmit={handleSubmit}>
-        <div className="mb-4">
-          <label htmlFor="name" className="form-label">Full Name</label>
-          <input
-            type="text"
-            id="name"
-            className="form-control form-control-lg shadow-sm"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            required
-            placeholder="Enter your full name"
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="email" className="form-label">Email Address</label>
-          <input
-            type="email"
-            id="email"
-            className="form-control form-control-lg shadow-sm"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-            placeholder="Enter your email"
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="message" className="form-label">Your Message</label>
-          <textarea
-            id="message"
-            className="form-control form-control-lg shadow-sm"
-            value={message}
-            onChange={(e) => setMessage(e.target.value)}
-            required
-            placeholder="Write your message here"
-            rows="5"
-          />
-        </div>
-        <div className="text-center">
-          <button type="submit" className="btn btn-lg btn-success shadow">Submit</button>
-        </div>
-      </form>
-    </div>
-  );
-};
-
-export default ContactUs;
